fix: activities-tabs-mode unnecessarily enabling tab-bar-mode when there are not tabs
`tab-bar-mode` automatically enables itself when a tab is created. There is no
need to display the default tab if the user did not use any tab command. This
way, the tab bar can be hidden until a tab/activity command is issued.

This is specially noticeable when the user has the following tab bar
configuration:
```
(setq tab-bar-close-last-tab-choice 'tab-bar-mode-disable)
```